Separates event manager out to another file.
