+++ lib/src/models/exports.dart
@@ -1,2 +1,3 @@
 export './list_item.dart' show ListItem;
+export './listbox_event_manager.dart' show ListboxEventManager;
 export './listbox_listener.dart' show ListboxListener;
