+++ lib/src/models/listbox_listener.dart
@@ -1,3 +1,4 @@
+/// Listen to events the Listbox emits
 mixin ListboxListener {
   void onListDragStart() {}
   void onListDragEnd() {}
