+++ lib/src/widgets/exports.dart
@@ -1,4 +1,4 @@
 export 'abstract_listbox_item.dart' show AbstractListboxItem;
 export 'simple_listbox_item.dart' show SimpleListboxItem;
 export 'templated_listbox_item.dart' show TemplatedListboxItem;
-export 'listbox.dart' show ListboxEventManager, Listbox;
+export 'listbox.dart' show Listbox;
